Add a scroll bar
This avoids the popover becoming too large. At the moment there
is no way to actually reach the scroll bar to scroll it, but it
prevents the popover from growing so large that it moves off
screen. The developer can then always navigate to the item in
question to view the detailed documentation.

Fixes #17