Reorganize some of the displayed information
This puts more relevant information, such as the summary and
descriptions of structural elements, up top. The "signature"
has been removed as the name is already clear to the user
as he is hovering on it when he wanted to see the tooltip.
Other information such as functiond and method parameter
lists were being displayed twice and were duplicating
information displayed by the call tips package.

Implements #13